+++ src/serviceworker/fetch/single-page-request.js
